+++ tests/pthread/test_pthread_barrier.cpp
@@ -21,7 +21,7 @@ int intermediate[N] = {};
 pthread_barrier_t barr;
 
 // Sums a single row of a matrix.
-int sum_row(int r)
+int sum_row(long r)
 {
     int sum = 0;
     for(int i = 0; i < N; ++i)
@@ -32,8 +32,8 @@ int sum_row(int r)
 void *thread_main(void *arg)
 {
     // Each thread sums individual rows.
-    int id = (int)arg;
-    for(int i = id; i < N; i += THREADS)
+    long id = (long)arg;
+    for(long i = id; i < N; i += THREADS)
         intermediate[i] = sum_row(i);
 
     // Synchronization point
@@ -69,7 +69,7 @@ int main(int argc, char **argv)
 
     // Barrier initialization
     int ret = pthread_barrier_init(&barr, NULL, THREADS);
-    assert(ret == 0); 
+    assert(ret == 0);
 
     for(int i = 0; i < THREADS; ++i) pthread_create(&thr[i], NULL, &thread_main, (void*)i);
     if (emscripten_has_threading_support())
